feat(core): add FieldValidation extension
